+++ src/ui/mod.rs
@@ -1,6 +1,6 @@
 #[allow(dead_code)] 
 pub mod util;
 pub mod app;
-pub mod repos_screen;    
+pub mod screen;
